requirements.txt
python-dotenv
fastapi
langchain
langchain-community
langchain-google-genai
langchain-groq
langgraph
pandas
SQLAlchemy
websockets
supabase==2.4.5
uvicorn
pymysql
python-multipart
gradio==4.20.0
langchain_openai